pytest: Access the right fields from the htlc_accepted payload
The plugin was attempting to access the field `req['htlc']['forward_amt']`
which does not exist. This being part of the payload for the processing
node is kept in the onion payload, not the htlc. And the field is called
`forward_amount` and is a string ending in `msat` as its unit. Before the
comparison that suffix needs to be stripped (and ensured to match `msat`)